generalisert stylingen for tjenesteboksene
