+++ src/App/Hovedside/Hovedside.tsx
@@ -2,11 +2,13 @@ import React, { FunctionComponent, useContext } from "react";
 
 import "./Hovedside.less";
 import TjenesteBoksContainer from "./TjenesteBoksContainer/TjenesteBoksContainer";
+import NyttigForDegContainer from "./NyttigForDegContainer/NyttigForDegContainer";
 
 const Hovedside: FunctionComponent = () => {
   return (
     <div className="forside">
       <TjenesteBoksContainer />
+      <NyttigForDegContainer />
     </div>
   );
 };
