+++ src/App/Hovedside/Innholdsboks/Innholdsboks.less
@@ -4,5 +4,4 @@
   background-color: white;
   border-radius: 4px;
   display: flex;
-  flex-grow: 1;
 }
