+++ src/App/Hovedside/NyttigForDegContainer/ArbeidsgiverTelefon/ArbeidsgiverTelefon.less
@@ -1,4 +1,4 @@
-@import '~nav-frontend-core/less/_variabler';
+@import '../../../../../node_modules/nav-frontend-core/less/_variabler';
 
 .arbeidsgivertelefon {
   height: 9rem;
