+++ src/App/Hovedside/NyttigForDegContainer/Bedriftsinfoknapp/Bedriftsinfoknapp.less
@@ -1,7 +1,7 @@
-@import '~nav-frontend-core/less/_variabler';
+@import '../../../../../node_modules/nav-frontend-core/less/_variabler';
 
-.kontakt-oss {
-  width: 23rem;
+.bedriftsinfoknapp {
+  width: 20rem;
   margin: 1rem;
 
   &__wrapper {
@@ -20,7 +20,7 @@
 }
 
 @media (min-width: @screen-sm-min) {
-  .kontakt-oss {
+  .bedriftsinfoknapp {
     width: 29rem;
     height: 8rem;
   }
