+++ src/App/Hovedside/NyttigForDegContainer/KontaktOss/KontaktOss.less
@@ -1,6 +1,6 @@
-@import '~nav-frontend-core/less/_variabler';
+@import '../../../../../node_modules/nav-frontend-core/less/_variabler';
 
-.bedriftsinfoknapp {
+.kontakt-oss {
   width: 23rem;
   margin: 1rem;
 
@@ -20,7 +20,7 @@
 }
 
 @media (min-width: @screen-sm-min) {
-  .bedriftsinfoknapp {
+  .kontakt-oss {
     width: 29rem;
     height: 8rem;
   }
