+++ src/App/Hovedside/NyttigForDegContainer/NyttigForDegContainer.less
@@ -1,0 +1,5 @@
+
+.nyttig-for-deg {
+  display: flex;
+  width: 66rem;
+}
