+++ src/App/Hovedside/TjenesteBoksContainer/Pamboks/Pamboks.less
@@ -3,6 +3,8 @@
 .pamboks {
   display: flex;
   flex-direction: column;
+  flex-grow: 1;
+
 
   &__rekruttering {
     margin-bottom: 0;
