+++ src/App/Hovedside/TjenesteBoksContainer/TjenesteBoksContainer.less
@@ -6,11 +6,11 @@
     flex-wrap: wrap;
 
     .tjenesteboks {
-        margin: 1rem;
         background-color: white;
         display:flex;
         flex-direction: column;
         height: 13.5rem;
+        margin: 1rem;
 
     }
